+++ Tests/HoneybadgerTests/main.m
@@ -9,6 +9,7 @@ void run_binary_image_tests(void);
 void run_signal_replay_tests(void);
 void run_exception_latch_tests(void);
 void run_signal_install_tests(void);
+void run_signal_chain_tests(void);
 
 int main(int argc, char** argv)
 {
@@ -21,6 +22,7 @@ int main(int argc, char** argv)
         run_signal_replay_tests();
         run_exception_latch_tests();
         run_signal_install_tests();
+        run_signal_chain_tests();
         return hb_test_summary();
     }
 }